move logic to SolanaRequestsLogic
